# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/locations/sites/{siteIDs}": {
            "get": {
                "description": "Retrieves locations for multiple sites",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "locations"
                ],
                "summary": "Get locations by site IDs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Comma-separated Site IDs",
                        "name": "siteIDs",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/sites": {
            "get": {
                "description": "Retrieves a paginated list of sites",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sites"
                ],
                "summary": "Get sites with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        },
...
        "/site/{siteID}": {
            "delete": {
                "description": "Deletes a site by its ID",
                "tags": [
                    "sites"
                ],
                "summary": "Delete site",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Site ID",
                        "name": "siteID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/CapIot-api_internal_models.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "CapIot-api_internal_models.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Use the new ErrorCode type",
                    "allOf": [
                        {
                            "$ref": "#/definitions/CapIot-api_internal_models.ErrorCode"
                        }
                    ]
                },
                "details": {
                    "description": "Optional: Additional details"
                },
                "message": {
                    "description": "Human-readable error message",
                    "type": "string"
                }
            }
        },
...
        "CapIot-api_internal_models.ErrorCode": {
            "type": "string",
            "enum": [
                "internal_server_error",
                "bad_request",
                "not_found",
                "forbidden",
                "unauthorized",
                "method_not_allowed",
                "invalid_token",
                "token_expired",
                "insufficient_permissions",
                "validation_failed",
                "missing_parameter",
                "invalid_format",
                "unsupported_media_type",
                "resource_not_found",
                "duplicate_resource"
            ],
            "x-enum-varnames": [
                "ErrorCodeInternalServerError",
                "ErrorCodeBadRequest",
                "ErrorCodeNotFound",
                "ErrorCodeForbidden",
                "ErrorCodeUnauthorized",
                "ErrorCodeMethodNotAllowed",
                "ErrorCodeInvalidToken",
                "ErrorCodeTokenExpired",
                "ErrorCodeInsufficientPermissions",
                "ErrorCodeValidationFailed",
                "ErrorCodeMissingParameter",
                "ErrorCodeInvalidFormat",
                "ErrorCodeUnsupportedMediaType",
                "ErrorCodeResourceNotFound",
                "ErrorCodeDuplicateResource"
            ]
        },
...
    }
}