# api/internal/handlers/location_handler.go
package handlers

import (
	"CapIot-api/internal/models"
	"CapIot-api/internal/service"
	"CapIot-api/internal/utils"
	"encoding/json"
	"log"
	"net/http"
	"strconv"
	"strings"

	"github.com/gorilla/mux"
)

type LocationHandler struct {
	locationService service.LocationService
}

func NewLocationHandler(locationService service.LocationService) *LocationHandler {
	return &LocationHandler{
		locationService: locationService,
	}
}

// CreateLocation godoc
// @Summary      Create a new location
// @Description  Creates a new location with the provided details
// @Tags         locations
// @Accept       json
// @Produce      json
// @Param        location body models.Location true "Location details"
// @Success      201  "Location created successfully"
// @Failure      400  {object}  models.APIError
// @Failure      415  {object}  models.APIError
// @Failure      500  {object}  models.APIError
// @Router       /api/location/create [post]
func (h *LocationHandler) CreateLocation(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		apiErr := models.NewAPIError(models.ErrorCodeMethodNotAllowed, "Method not allowed", nil, http.StatusMethodNotAllowed)
		utils.RespondWithError(w, apiErr)
		return
	}

	// Ensure Content-Type is application/json
	if r.Header.Get("Content-Type") != "application/json" {
		apiErr := models.NewAPIError(models.ErrorCodeUnsupportedMediaType, "Unsupported Content-Type", nil, http.StatusUnsupportedMediaType)
		utils.RespondWithError(w, apiErr)
		return
	}

	// Decode JSON body into Location struct
	var location models.Location
	if err := json.NewDecoder(r.Body).Decode(&location); err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Invalid JSON", map[string]string{"error": err.Error()}, http.StatusBadRequest)
		utils.RespondWithError(w, apiErr)
		return
	}

	// Validate input
	if location.Name == nil || *location.Name == "" {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Location name is required", nil, http.StatusBadRequest)
		utils.RespondWithError(w, apiErr)
		return
	}
	if location.Description == nil {
		empty := ""
		location.Description = &empty
	}

	log.Printf("Creating location: %+v\n", location)

	// Call service to create location
	if err := h.locationService.CreateLocation(location); err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeInternalServerError, "Error creating location", map[string]string{"error": err.Error()}, http.StatusInternalServerError)
		utils.RespondWithError(w, apiErr)
		return
	}

	w.WriteHeader(http.StatusCreated)
	w.Write([]byte("Location created successfully"))
}

// GetComponentsBylocationID godoc
// @Summary      Get components by location ID
// @Description  Retrieves all components associated with a specific location
// @Tags         locations
// @Produce      json
// @Param        locationID path int true "Location ID"
// @Success      200  {array}   models.Component
// @Failure      400  {object}  models.APIError
// @Failure      500  {object}  models.APIError
// @Router       /api/locations/{locationID}/components [get]
func (h *LocationHandler) GetComponentsBylocationID(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		apiErr := models.NewAPIError(models.ErrorCodeMethodNotAllowed, "Method not allowed", nil, http.StatusMethodNotAllowed)
		utils.RespondWithError(w, apiErr)
		return
	}

	vars := mux.Vars(r)
	locationID, ok := vars["locationID"]
	if !ok {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Missing location ID in path", nil, http.StatusBadRequest)
		utils.RespondWithError(w, apiErr)
		return
	}

	components, err := h.locationService.GetComponentsBylocationID(locationID)
	if err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeInternalServerError, "Error getting components", map[string]string{"error": err.Error()}, http.StatusInternalServerError)
		utils.RespondWithError(w, apiErr)
		return
	}

	utils.RespondWithJSON(w, http.StatusOK, components)
}

// GetDevicesBylocationID godoc
// @Summary      Get devices by location ID
// @Description  Retrieves all devices associated with a specific location
// @Tags         locations
// @Produce      json
// @Param        locationID path int true "Location ID"
// @Success      200  {array}   models.Device
// @Failure      400  {object}  models.APIError
// @Failure      500  {object}  models.APIError
// @Router       /api/location/{locationID}/devices [get]
func (h *LocationHandler) GetDevicesBylocationID(writer http.ResponseWriter, request *http.Request) {
	if request.Method != http.MethodGet {
		apiErr := models.NewAPIError(models.ErrorCodeMethodNotAllowed, "Method not allowed", nil, http.StatusMethodNotAllowed)
		utils.RespondWithError(writer, apiErr)
		return
	}

	vars := mux.Vars(request)
	locationID, ok := vars["locationID"]
	if !ok {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Missing location ID in path", nil, http.StatusBadRequest)
		utils.RespondWithError(writer, apiErr)
		return
	}

	devices, err := h.locationService.GetDevicesBylocationID(locationID)
	if err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeInternalServerError, "Error getting devices", map[string]string{"error": err.Error()}, http.StatusInternalServerError)
		utils.RespondWithError(writer, apiErr)
		return
	}

	utils.RespondWithJSON(writer, http.StatusOK, devices)
}

// GetAllLocations godoc
// @Summary      Get all locations
// @Description  Retrieves a paginated list of all locations
// @Tags         locations
// @Produce      json
// @Param        page query int false "Page number"
// @Param        limit query int false "Items per page"
// @Param        search query string false "Search term"
// @Success      200  {object}  map[string]interface{}
// @Failure      500  {object}  models.APIError
// @Router       /api/locations [get]
func (h *LocationHandler) GetAllLocations(writer http.ResponseWriter, request *http.Request) {
	if request.Method != http.MethodGet {
		apiErr := models.NewAPIError(models.ErrorCodeMethodNotAllowed, "Method not allowed", nil, http.StatusMethodNotAllowed)
		utils.RespondWithError(writer, apiErr)
		return
	}

	// Get pagination parameters from query string
	pageStr := request.URL.Query().Get("page")
	limitStr := request.URL.Query().Get("limit")

	page := 1
	limit := 10 // Default page size

	if pageStr != "" {
		if p, err := strconv.Atoi(pageStr); err == nil && p > 0 {
			page = p
		}
	}

	if limitStr != "" {
		if l, err := strconv.Atoi(limitStr); err == nil && l > 0 && l <= 100 { // Add a reasonable max limit
			limit = l
		}
	}

	// Get the optional search parameter
	searchTerm := request.URL.Query().Get("search")
	// Call AuthService to get all locations
	locations, err := h.locationService.GetAllLocations(request.Context(), page, limit, searchTerm)
	if err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeInternalServerError, "Failed to retrieve data", map[string]string{"error": err.Error()}, http.StatusInternalServerError)
		utils.RespondWithError(writer, apiErr)
		return
	}

	utils.RespondWithJSON(writer, http.StatusOK, locations)
}

// CreateSite godoc
// @Summary      Create a new site
// @Description  Creates a new site with the provided details
// @Tags         sites
// @Accept       json
// @Produce      json
// @Param        site body models.Site true "Site details"
// @Success      201  {object}  models.Site
// @Failure      400  {object}  models.APIError
// @Failure      415  {object}  models.APIError
// @Failure      500  {object}  models.APIError
// @Router       /site/create [post]
func (h *LocationHandler) CreateSite(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		apiErr := models.NewAPIError(models.ErrorCodeMethodNotAllowed, "Method not allowed", nil, http.StatusMethodNotAllowed)
		utils.RespondWithError(w, apiErr)
		return
	}

	if r.Header.Get("Content-Type") != "application/json" {
		apiErr := models.NewAPIError(models.ErrorCodeUnsupportedMediaType, "Unsupported Content-Type", nil, http.StatusUnsupportedMediaType)
		utils.RespondWithError(w, apiErr)
		return
	}

	var site models.Site
	if err := json.NewDecoder(r.Body).Decode(&site); err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Invalid JSON", map[string]string{"error": err.Error()}, http.StatusBadRequest)
		utils.RespondWithError(w, apiErr)
		return
	}

	// Validate input
	if site.Name == nil || *site.Name == "" {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Site name is required", nil, http.StatusBadRequest)
		utils.RespondWithError(w, apiErr)
		return
	}

	log.Printf("Creating site: %+v\n", site)

	// Call service to create site
	createdSite, err := h.locationService.CreateSite(r.Context(), site)
	if err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeInternalServerError, "Error creating site", map[string]string{"error": err.Error()}, http.StatusInternalServerError)
		utils.RespondWithError(w, apiErr)
		return
	}

	utils.RespondWithJSON(w, http.StatusCreated, createdSite)
}

// DeleteSite godoc
// @Summary      Delete site
// @Description  Deletes a site by its ID
// @Tags         sites
// @Param        siteID path int true "Site ID"
// @Success      204  "No Content"
// @Failure      400  {object}  models.APIError
// @Failure      405  {object}  models.APIError
// @Failure      500  {object}  models.APIError
// @Router       /site/{siteID} [delete]
func (h *LocationHandler) DeleteSite(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodDelete {
		apiErr := models.NewAPIError(models.ErrorCodeMethodNotAllowed, "Method not allowed", nil, http.StatusMethodNotAllowed)
		utils.RespondWithError(w, apiErr)
		return
	}

	vars := mux.Vars(r)
	idStr, ok := vars["siteID"]
	if !ok {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Missing site ID in path", nil, http.StatusBadRequest)
		utils.RespondWithError(w, apiErr)
		return
	}

	id, err := strconv.ParseInt(idStr, 10, 64)
	if err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Invalid site ID", map[string]string{"error": err.Error()}, http.StatusBadRequest)
		utils.RespondWithError(w, apiErr)
		return
	}

	if err := h.locationService.DeleteSite(r.Context(), id); err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeInternalServerError, "Error deleting site", map[string]string{"error": err.Error()}, http.StatusInternalServerError)
		utils.RespondWithError(w, apiErr)
		return
	}

	w.WriteHeader(http.StatusNoContent) // 204 No Content
}

// GetSitesWithPagination godoc
// @Summary      Get sites with pagination
// @Description  Retrieves a paginated list of sites
// @Tags         sites
// @Produce      json
// @Param        page query int false "Page number"
// @Param        limit query int false "Items per page"
// @Param        search query string false "Search term"
// @Success      200  {object}  map[string]interface{}
// @Failure      405  {object}  models.APIError
// @Failure      500  {object}  models.APIError
// @Router       /api/sites [get]
func (h *LocationHandler) GetSitesWithPagination(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		apiErr := models.NewAPIError(models.ErrorCodeMethodNotAllowed, "Method not allowed", nil, http.StatusMethodNotAllowed)
		utils.RespondWithError(w, apiErr)
		return
	}

	// Get pagination parameters from query string
	pageStr := r.URL.Query().Get("page")
	limitStr := r.URL.Query().Get("limit")

	page := 1
	limit := 10 // Default page size

	if pageStr != "" {
		if p, err := strconv.Atoi(pageStr); err == nil && p > 0 {
			page = p
		}
	}

	if limitStr != "" {
		if l, err := strconv.Atoi(limitStr); err == nil && l > 0 && l <= 100 { // Add a reasonable max limit
			limit = l
		}
	}

	// Get the optional search parameter
	searchTerm := r.URL.Query().Get("search")

	sitesData, err := h.locationService.GetSitesWithPagination(r.Context(), page, limit, searchTerm)
	if err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeInternalServerError, "Failed to retrieve data", map[string]string{"error": err.Error()}, http.StatusInternalServerError)
		utils.RespondWithError(w, apiErr)
		return
	}

	utils.RespondWithJSON(w, http.StatusOK, sitesData)
}

// GetLocationsBySiteIDs godoc
// @Summary      Get locations by site IDs
// @Description  Retrieves locations for multiple sites
// @Tags         locations
// @Produce      json
// @Param        siteIDs path string true "Comma-separated Site IDs"
// @Param        page query int false "Page number"
// @Param        limit query int false "Items per page"
// @Param        search query string false "Search term"
// @Success      200  {object}  map[string]interface{}
// @Failure      400  {object}  models.APIError
// @Failure      405  {object}  models.APIError
// @Failure      500  {object}  models.APIError
// @Router       /api/locations/sites/{siteIDs} [get]
func (h *LocationHandler) GetLocationsBySiteIDs(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodGet {
		apiErr := models.NewAPIError(models.ErrorCodeMethodNotAllowed, "Method not allowed", nil, http.StatusMethodNotAllowed)
		utils.RespondWithError(w, apiErr)
		return
	}

	// Read query params
	query := r.URL.Query()
	// Renamed the path variable to siteIDParam to avoid confusion with the slice of IDs
	siteIDParam := mux.Vars(r)["siteIDs"]
	if siteIDParam == "" {
		apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Missing site IDs in path", nil, http.StatusBadRequest)
		utils.RespondWithError(w, apiErr)
		return
	}

	// Read page & limit for pagination (optional)
	pageStr := query.Get("page")
	limitStr := query.Get("limit")
	page, _ := strconv.Atoi(pageStr)
	if page < 1 {
		page = 1
	}
	limit, _ := strconv.Atoi(limitStr)
	if limit < 1 {
		limit = 10 // default page size
	}

	// Read search term (optional)
	term := query.Get("search")
	// Default to empty string if not provided
	if term == "" {
		term = ""
	}

	// Convert comma-separated siteIDParam to slice of ints
	idStrs := strings.Split(siteIDParam, ",")
	var siteIDInts []int
	for _, rawIDStr := range idStrs {
		// Trim surrounding whitespace from each ID string
		idStr := strings.TrimSpace(rawIDStr)

		if idStr == "" {
			continue // Skip empty strings that result from splitting, e.g., "4,,5" or trailing commas
		}

		id, err := strconv.Atoi(idStr)
		if err != nil {
			// IMPORTANT: This error is triggered when non-integer values are in the comma-separated list
			apiErr := models.NewAPIError(models.ErrorCodeBadRequest, "Invalid site ID format. IDs must be integers.", map[string]string{"error": err.Error()}, http.StatusBadRequest)
			utils.RespondWithError(w, apiErr)
			return
		}
		siteIDInts = append(siteIDInts, id)
	}

	// Call service with the validated slice of integers (siteIDInts)
	// NOTE: The service function signature must be updated to accept []int instead of string.
	locations, err := h.locationService.GetLocationsBySiteIDs(r.Context(), siteIDInts, page, limit, term)
	if err != nil {
		apiErr := models.NewAPIError(models.ErrorCodeInternalServerError, "Error getting locations", map[string]string{"error": err.Error()}, http.StatusInternalServerError)
		utils.RespondWithError(w, apiErr)
		return
	}

	utils.RespondWithJSON(w, http.StatusOK, locations)
}

func (h *LocationHandler) CheckSiteAccess(id int, idInt int64) (bool, error) {
	return h.locationService.CheckUserAccessToSite(id, idInt)
}

func (h *LocationHandler) CheckLocationAccess(id int, id2 int64) (bool, error) {
	return h.locationService.CheckUserAccessToLocation(id, id2)
}